OPENJPA-869 Fixup order column test variations failing on Oracle.
git-svn-id: https://svn.apache.org/repos/asf/openjpa/trunk@813556 13f79535-47bb-0310-9956-ffa450edef68

+++ openjpa-persistence-jdbc/src/test/java/org/apache/openjpa/persistence/jdbc/order/ColDefTestEntity.java
@@ -43,15 +43,15 @@ public class ColDefTestEntity {
     private int id;
     
     @OneToMany(cascade=CascadeType.PERSIST)
-    @OrderColumn(name="one2McoldefOrder", columnDefinition="BIGINT")
+    @OrderColumn(name="one2McoldefOrder", columnDefinition="INTEGER")
     private List<ColDefTestElement> one2Mcoldef;
 
     @ElementCollection
-    @OrderColumn(name="colcoldefOrder", columnDefinition="BIGINT")
+    @OrderColumn(name="colcoldefOrder", columnDefinition="INTEGER")
     private Set<ColDefTestElement> collcoldef;
     
     @ManyToMany(cascade=CascadeType.PERSIST)
-    @OrderColumn(name="m2McoldefOrder", columnDefinition="BIGINT")
+    @OrderColumn(name="m2McoldefOrder", columnDefinition="INTEGER")
     private List<ColDefTestElement> m2mcoldef;
 
     public void setOne2Mcoldef(List<ColDefTestElement> one2Mcoldef) {
